Blur Faces
have to change some of the paths (idk why mine only works with my full file path),. im not sure if we are allowed to use other face_detector xml because teacher gave it in the original folder, and didnt specify in the question pdf. but the hints he gave specifies he used the face detector xml he gave so i just used it, and so the face detection isnt as good and similar to the sample vid. might have to ask tcher about this.